Add createGrid() & center #container
createGrid() will take a number and append .box
element to .row the amount of that number given.
So it will something like boxes = box * number,
which will then be appended to an individual row.
Then the function will make copies of that row
and append it to the #container element the same
number of times. So it will look like: -- below
--> rows = row * number. Thus, if 16 is passed
into createGrid(), there will be a grid of 16x16
boxes all attached to the #container.